+++ src/cqindustry/walkway/__init__.py
@@ -1,1 +1,0 @@
-from .Walkway import Walkway